Fix incorrect error message + split tests
